Fix merge conflicts after rebase
